Updated styles, refactored, added Intendity editor
Updated form styles for light/dark mode to increase contrast & make them a bit more interesting.

Reworked the value range for NormalizedSlider and the way the Rotation editor converts between that and radians. I was trying to get Safari to have more precision when updating values as the slider changes, but it looks like there's just no way to do that; Safari seems to have a pretty coarse level of granularity for slider change deltas.

I did this because, in Safari, the smallest interval you can change a rotation axis by is 3 degrees. I wanted to get it down to one. That isn't possible. I confirmed Rotation changes by < one degree in Firefox (the display value occasionally skips a number due to rounding). I don't see any way to get things to work better in Safari :-(