Flipped the selector pane to the left side of the screen
Initially, I had ther selector pane on the right because I wanted to PrimaryWindow to be, well, most prominent. I've been thinking about it, though, and at some point Amber will support both node-based and text-based programming. The text-based programming will likely be implemented using VSCode - which has it's manu pane on the left ... just as every editor/IDE I can think of does.

So, to be more in line with this convention, the selector pane was moved to the left side of the screen 😊