Major refactor to creators, added toggleable geometry wireframes
Instead of being a component, a creator is now just a function. The only ui for entities is now the editors. The Library component was refactored so that whenever a Creator item in its list is selected the appropriate creator function executes.

Entities can also be renamed now. This is a requirement for how entities are created/edited now. When created, an entity gets the default name ':placeholder:'. This can then be changed. This changing is possible because creators no longer specify the id of the entity they create. Instead, they link the editor id (:placeholder: by default) to the auto-generated ecs id. With this mapping, editor ids can be changed easily (as their mapping to ecs id is maintained).