Updated how light rotation works
Lights are no longer rotated by moving their targets. Instead, they group they and their target are added to is rotated. This works because, when added to a group, the position of both the light and the target become relative to the group's position. This means that when the group position or rotation are changed those changes propagate to the target, and the light.

This means that the LightRotation editor is no longer needed as lights can be rotated using the Rotation editor just like meshes.