Style changes, Color & Object track expanded state
Still working on the color scheme. Not sure I've found what I want to settle on yet.

Did some initial work on styling for the library selector. I'm not using variables yet, as I'm not settled enough on the styles for that.

A major UX win is that both the Object and Color ui components now keep track of their expanded/collapsed state per property and per entity; so as you switch between entities in the library selector, component editors will stay expanded/collapsed. This was surprisingly easy to implement and is such a huge usability win.